big refactor to get tests to work properly in packaged mode. moved MockRequest class into new test_utils.py
