+++ juriscraper/__init__.py
@@ -2,3 +2,15 @@ __all__ = [
     'opinions',
     'oral_args',
 ]
+
+__version__ = "1.1"
+
+__title__ = "juriscraper"
+__description__ = "An API to scrape American court websites for metadata."
+__uri__ = "https://github.com/freelawproject/juriscraper"
+
+__author__ = "Mike Lissner"
+__email__ = "info@freelawproject.org"
+
+__license__ = "BSD"
+__copyright__ = "Copyright (c) Free Law Project"
